+++ code75k0.c
@@ -275,14 +275,14 @@ chk:
  */
 
 /*!------------------------------------------------------------------------
- * \fn     DissectBit_75K0(char *pDest, int DestSize, LargeWord Inp)
+ * \fn     DissectBit_75K0(char *pDest, size_t DestSize, LargeWord Inp)
  * \brief  dissect compact storage of bit (field) into readable form for listing
  * \param  pDest destination for ASCII representation
  * \param  DestSize destination buffer size
  * \param  Inp compact storage
  * ------------------------------------------------------------------------ */
 
-static void DissectBit_75K0(char *pDest, int DestSize, LargeWord Inp)
+static void DissectBit_75K0(char *pDest, size_t DestSize, LargeWord Inp)
 {
   if (Hi(Inp))
     as_snprintf(pDest, DestSize, "%~03.*u%s.%c",
