+++ codem16.c
@@ -181,7 +181,7 @@ static Boolean DecodeRegCore(const char *pArg, Word *pResult)
 }
 
 /*!------------------------------------------------------------------------
- * \fn     DissectReg_M16(char *pDest, int DestSize, tRegInt Value, tSymbolSize InpSize)
+ * \fn     DissectReg_M16(char *pDest, size_t DestSize, tRegInt Value, tSymbolSize InpSize)
  * \brief  dissect register symbols - M16 variant
  * \param  pDest destination buffer
  * \param  DestSize destination buffer size
@@ -189,7 +189,7 @@ static Boolean DecodeRegCore(const char *pArg, Word *pResult)
  * \param  InpSize register size
  * ------------------------------------------------------------------------ */
 
-static void DissectReg_M16(char *pDest, int DestSize, tRegInt Value, tSymbolSize InpSize)
+static void DissectReg_M16(char *pDest, size_t DestSize, tRegInt Value, tSymbolSize InpSize)
 {
   if (InpSize == eSymbolSize32Bit)
   {
