+++ codemcore.c
@@ -103,7 +103,7 @@ static Boolean DecodeRegCore(const char *pArg, Word *pResult)
 }
 
 /*!------------------------------------------------------------------------
- * \fn     DissectReg_MCORE(char *pDest, int DestSize, tRegInt Value, tSymbolSize InpSize)
+ * \fn     DissectReg_MCORE(char *pDest, size_t DestSize, tRegInt Value, tSymbolSize InpSize)
  * \brief  dissect register symbols - M-CORE variant
  * \param  pDest destination buffer
  * \param  DestSize destination buffer size
@@ -111,7 +111,7 @@ static Boolean DecodeRegCore(const char *pArg, Word *pResult)
  * \param  InpSize register size
  * ------------------------------------------------------------------------ */
 
-static void DissectReg_MCORE(char *pDest, int DestSize, tRegInt Value, tSymbolSize InpSize)
+static void DissectReg_MCORE(char *pDest, size_t DestSize, tRegInt Value, tSymbolSize InpSize)
 {
   switch (InpSize)
   {
