+++ codest6.c
@@ -250,14 +250,14 @@ static Boolean DissectBitSymbol(LongWord BitSymbol, Word *pAddress, Byte *pBitPo
 }
 
 /*!------------------------------------------------------------------------
- * \fn     DissectBit_ST6(char *pDest, int DestSize, LargeWord Inp)
+ * \fn     DissectBit_ST6(char *pDest, size_t DestSize, LargeWord Inp)
  * \brief  dissect compact storage of bit (field) into readable form for listing
  * \param  pDest destination for ASCII representation
  * \param  DestSize destination buffer size
  * \param  Inp compact storage
  * ------------------------------------------------------------------------ */
 
-static void DissectBit_ST6(char *pDest, int DestSize, LargeWord Inp)
+static void DissectBit_ST6(char *pDest, size_t DestSize, LargeWord Inp)
 {
   Byte BitPos;
   Word Address;
