+++ codesx20.c
@@ -98,14 +98,14 @@ static void SplitBitSymbol(Word BitSymbol, Word *pRegAddr, Word *pBitPos)
 }
 
 /*!------------------------------------------------------------------------
- * \fn     DissectBit_SX20(char *pDest, int DestSize, LargeWord Inp)
+ * \fn     DissectBit_SX20(char *pDest, size_t DestSize, LargeWord Inp)
  * \brief  dissect compact storage of bit into readable form for listing
  * \param  pDest destination for ASCII representation
  * \param  DestSize destination buffer size
  * \param  Inp compact storage
  * ------------------------------------------------------------------------ */
 
-static void DissectBit_SX20(char *pDest, int DestSize, LargeWord Inp)
+static void DissectBit_SX20(char *pDest, size_t DestSize, LargeWord Inp)
 {
   Word BitPos, Address;
