+++ tex2doc.c
@@ -957,7 +957,7 @@ static void DoAddNormal(const char *Part, char *Sep)
   }
 }
 
-static void GetTableName(char *Dest, int DestSize)
+static void GetTableName(char *Dest, size_t DestSize)
 {
   int ThisTableNum = (CurrEnv == EnvTabular) ? TableNum + 1 : TableNum;
 
@@ -967,7 +967,7 @@ static void GetTableName(char *Dest, int DestSize)
     as_snprintf(Dest, DestSize, "%d.%d", Chapters[0], ThisTableNum);
 }
 
-static void GetSectionName(char *Dest, int DestSize)
+static void GetSectionName(char *Dest, size_t DestSize)
 {
   int z;
